% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_acc.R
\name{horse_acc}
\alias{horse_acc}
\title{Horse Acceleration}
\usage{
horse_acc(data)
}
\description{
This function calculates the radial acceleration in meters per squaresecond
from gps velocity (filtered) for multiple rides in the same dataframe
The dataframe "data" has to be structured as follows:
col "ID" with horse identifier, col "Zeit" with timestamp,
col "Delta_t" with time difference in seconds,
col "V_gefiltert" with filtered velocity in meters per minute
}
