% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_COT.R
\name{horse_COT}
\alias{horse_COT}
\title{Horse Oxygen Cost of Transport}
\usage{
horse_COT(data)
}
\description{
This function calculates the Oxygen Cost of Transport (COT in ml/kg)
from gps distance, acceleration and elevation for multiple rides in the same dataframe
The dataframe "data" has to be structured as follows:
col "ID" with horse identifier, col "Zeit" with timestamp (tz = "Europe/Berlin"),
col "GPSDistance" with gps distance in m
col "Acceleration" with radial acceleration in m/s^2
col "Altitude" with elevation in m
}
