% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_filter.R
\name{horse_filter}
\alias{horse_filter}
\title{Horse Filter}
\usage{
horse_filter(data)
}
\description{
This function filters the velocity and heart rate
for multiple rides in the same dataframe
The dataframe "data" has to be structured as follows:
col "ID" with horse identifier, col "Zeit" with timestamp,
col "Lat" with Latitude (decimal), col "Lon" with Longitude (decimal),
col "HF" with heart rate in bpm, col "V" with velocity in m/min
}
