% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_findIntervals.R
\name{horse_findIntervals}
\alias{horse_findIntervals}
\title{Horse Find Intervals}
\usage{
horse_findIntervals(data)
}
\description{
This function assigns intervals to sections of the data
for multiple rides in the same dataframe
The dataframe "data" has to be structured as follows:
col "ID" with horse identifier, col "Zeit" with timestamp (tz = "Europe/Berlin"),
col "V" with velocity in m/min, col "Gait" with gait (1 = walk, 2 = trot, 3 = gallop)
}
