% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_GPSDist.R
\name{horse_GPSDist}
\alias{horse_GPSDist}
\title{Horse GPS Distance}
\usage{
horse_GPSDist(data)
}
\description{
This function calculates the distance (GPSDistance) in meters between locations
from gps latitude (Lat) and longitude (Lon) for multiple rides in the same dataframe
The dataframe "data" has to be structured as follows:
col "ID" with horse identifier, col "Zeit" with timestamp (tz = "Europe/Berlin"),
col "Lat" with Latitude (decimal), col "Lon" with Longitude (decimal)
}
