% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horse_heightProfilefromXCApp.R
\name{horse_heightProfilefromXCApp}
\alias{horse_heightProfilefromXCApp}
\title{Make height profile from Cross-Country App}
\usage{
horse_heightProfilefromXCApp(kml, efforts)
}
\description{
This function uses data from Cross-Country App and creates
a height profile
Following objects must be present:
"kml": kml-file with track and efforts from Cross-Country App
list with length = number of efforts + number of minute markers + track,
efforts and minute markers: double(1 x Lon, Lat, Alt),
track: double(length track x Lon, Lat, Alt)
"efforts": csv-file with efforts from Cross-Country App
dataframe (Type, Number, Name, Strides, Seconds.From.Start(mm:ss),
Metres.From.Start, Seconds.To.Finish(mm:ss), Metres.To.Finish,
Seconds.From.Previous(mm:ss), Metres.From.Previous, Comment)
}
